free almost ok

+++ src/history/history_option2.c
@@ -6,27 +6,19 @@
 /*   By: cboussau <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/07/17 15:11:34 by cboussau          #+#    #+#             */
-/*   Updated: 2016/10/22 11:28:42 by qdiaz            ###   ########.fr       */
+/*   Updated: 2016/10/22 18:07:41 by qdiaz            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <sh42.h>
 
-static void		add_to_file(t_hub *info)
+static void		add_to_file_bis(t_hub *info, int fd)
 {
 	t_dlist	*dlst;
-	char	*pathb;
-	int		fd;
 	int		i;
 
 	i = 1;
 	dlst = info->node;
-	pathb = ft_strjoin("/tmp/history", "bis");
-	if ((fd = open(pathb, O_RDWR | O_CREAT, 0644)) == -1)
-	{
-		ft_putendl_fd("history : No such file or directory", 2);
-		exit(-1);
-	}
 	while (info->node->prev)
 		info->node = info->node->prev;
 	while (info->node)
@@ -37,8 +29,24 @@ static void		add_to_file(t_hub *info)
 		i++;
 	}
 	info->node = dlst;
+}
+
+static void		add_to_file(t_hub *info)
+{
+	char	*pathb;
+	int		fd;
+
+	pathb = ft_strjoin("/tmp/history", "bis");
+	if ((fd = open(pathb, O_RDWR | O_CREAT, 0644)) == -1)
+	{
+		ft_putendl_fd("history : No such file or directory", 2);
+		free(pathb);
+		return ;
+	}
+	add_to_file_bis(info, fd);
 	unlink("/tmp/history");
 	rename(pathb, "/tmp/history");
+	free(pathb);
 }
 
 void			option_r(t_hub *info)
@@ -49,7 +57,7 @@ void			option_r(t_hub *info)
 	if ((fd = open("/tmp/history", O_RDWR | O_CREAT, 0644)) == -1)
 	{
 		ft_putendl_fd("history : No such file or directory", 2);
-		exit(-1);
+		return ;
 	}
 	while (get_next_line(fd, &line) > 0)
 	{
